refactor(makeExecutableSchema): add subschema support
i.e. unify with stitchSchemas
