+++ src/generate/concatenateTypeDefs.ts
@@ -24,7 +24,7 @@ function concatenateTypeDefs(
     } else {
       const type = typeof typeDef;
       throw new SchemaError(
-        `typeDef array must contain only strings and functions, got ${type}`,
+        `typeDef array must contain only strings, documents, or functions, got ${type}`,
       );
     }
   });
