+++ src/generate/mergeInfo.ts
