+++ src/generate/typeFromAST.ts
@@ -27,11 +27,13 @@ import {
   GraphQLEnumValueConfigMap,
   GraphQLFieldConfigArgumentMap,
   valueFromASTUntyped,
+  EnumValueDefinitionNode,
+  getDirectiveValues,
+  GraphQLDeprecatedDirective,
 } from 'graphql';
 
 import { graphqlVersion } from '../utils/index';
 import { createStub, createNamedStub } from '../utils/stub';
-import resolveFromParentTypename from '../utils/resolveFromParentTypename';
 
 const backcompatOptions = { commentDescriptions: true };
 
@@ -81,7 +83,7 @@ function makeObjectType(node: ObjectTypeDefinitionNode): GraphQLObjectType {
 function makeInterfaceType(
   node: InterfaceTypeDefinitionNode,
 ): GraphQLInterfaceType {
-  const config = {
+  return new GraphQLInterfaceType({
     name: node.name.value,
     description: getDescription(node, backcompatOptions),
     interfaces:
@@ -92,10 +94,8 @@ function makeInterfaceType(
             )
         : undefined,
     fields: () => makeFields(node.fields),
-    resolveType: (parent: any) => resolveFromParentTypename(parent),
     astNode: node,
-  };
-  return new GraphQLInterfaceType(config);
+  });
 }
 
 function makeEnumType(node: EnumTypeDefinitionNode): GraphQLEnumType {
@@ -104,6 +104,8 @@ function makeEnumType(node: EnumTypeDefinitionNode): GraphQLEnumType {
       ...prev,
       [value.name.value]: {
         description: getDescription(value, backcompatOptions),
+        deprecationReason: getDeprecationReason(value),
+        astNode: value,
       },
     }),
     {},
@@ -123,7 +125,6 @@ function makeUnionType(node: UnionTypeDefinitionNode): GraphQLUnionType {
     description: getDescription(node, backcompatOptions),
     types: () =>
       node.types.map((type) => createNamedStub(type.name.value, 'object')),
-    resolveType: (parent) => resolveFromParentTypename(parent),
     astNode: node,
   });
 }
@@ -132,13 +133,9 @@ function makeScalarType(node: ScalarTypeDefinitionNode): GraphQLScalarType {
   return new GraphQLScalarType({
     name: node.name.value,
     description: getDescription(node, backcompatOptions),
-    serialize: () => null,
-    // Note: validation calls the parse functions to determine if a
-    // literal value is correct. Returning null would cause use of custom
-    // scalars to always fail validation. Returning false causes them to
-    // always pass validation.
-    parseValue: () => false,
-    parseLiteral: () => false,
+    serialize: (value) => value,
+    parseValue: (value) => value,
+    parseLiteral: (value) => value,
     astNode: node,
   });
 }
@@ -157,31 +154,19 @@ function makeInputObjectType(
 function makeFields(
   nodes: ReadonlyArray<FieldDefinitionNode>,
 ): Record<string, GraphQLFieldConfig<any, any>> {
-  return nodes.reduce((prev, node) => {
-    const deprecatedDirective = node.directives.find(
-      (directive) => directive.name.value === 'deprecated',
-    );
-
-    let deprecationReason;
-
-    if (deprecatedDirective != null) {
-      const deprecatedArgument = deprecatedDirective.arguments.find(
-        (arg) => arg.name.value === 'reason',
-      );
-      deprecationReason = (deprecatedArgument.value as StringValueNode).value;
-    }
-
-    return {
+  return nodes.reduce(
+    (prev, node) => ({
       ...prev,
       [node.name.value]: {
         type: createStub(node.type, 'output'),
         description: getDescription(node, backcompatOptions),
         args: makeValues(node.arguments),
-        deprecationReason,
+        deprecationReason: getDeprecationReason(node),
         astNode: node,
       },
-    };
-  }, {});
+    }),
+    {},
+  );
 }
 
 function makeValues(
@@ -323,4 +308,11 @@ function leadingWhitespace(str: string) {
 
 function isBlank(str: string) {
   return leadingWhitespace(str) === str.length;
+}
+
+function getDeprecationReason(
+  node: EnumValueDefinitionNode | FieldDefinitionNode,
+): string {
+  const deprecated = getDirectiveValues(GraphQLDeprecatedDirective, node);
+  return deprecated?.reason;
 }
