+++ src/index.ts
@@ -7,6 +7,5 @@ export * from './links';
 export * from './mock';
 export * from './polyfills';
 export * from './scalars';
-export * from './stitch';
 export * from './wrap';
 export * from './utils';
