--- src/stitch/index.ts
@@ -1,4 +1,0 @@
-export { default as stitchSchemas } from './stitchSchemas';
-
-// for backwards compatibility
-export { default as mergeSchemas } from './stitchSchemas';
