+++ src/test/extensionExtraction.test.ts
@@ -1,6 +1,6 @@
 import { parse } from 'graphql';
 
-import { extractExtensionDefinitions } from '../generate/extensionDefinitions';
+import { extractExtensionDefinitions } from '../generate/definitions';
 
 describe('Extension extraction', () => {
   test('extracts extended inputs', () => {
