+++ src/test/schemaGenerator.test.ts
@@ -100,12 +100,6 @@ describe('generating schema from shorthand', () => {
     expect(() => makeExecutableSchema(undefined)).toThrowError('undefined');
   });
 
-  test('throws an error if typeDefinitionNodes are not provided', () => {
-    expect(() =>
-      makeExecutableSchema({ typeDefs: undefined, resolvers: {} }),
-    ).toThrowError('Must provide typeDefs');
-  });
-
   test('throws an error if no resolveFunctions are provided', () => {
     expect(() =>
       makeExecutableSchema({ typeDefs: 'blah', resolvers: {} }),
@@ -130,7 +124,7 @@ describe('generating schema from shorthand', () => {
         resolvers: {},
       }),
     ).toThrowError(
-      'typeDef array must contain only strings and functions, got number',
+      'typeDef array must contain only strings, documents, or functions, got number',
     );
   });
