+++ src/test/stitchSchemas.test.ts
@@ -15,7 +15,6 @@ import {
 import { delegateToSchema } from '../delegate/index';
 import { makeExecutableSchema } from '../generate/index';
 import { IResolvers, SubschemaConfig } from '../Interfaces';
-import { stitchSchemas } from '../stitch/index';
 import {
   cloneSchema,
   getResolversFromSchema,
@@ -344,17 +343,19 @@ testCombinations.forEach((combination) => {
       bookingSchema = await combination.booking;
       productSchema = await combination.product;
 
-      stitchedSchema = stitchSchemas({
-        schemas: [
+      stitchedSchema = makeExecutableSchema({
+        subschemas: [
           propertySchema,
           bookingSchema,
           productSchema,
-          interfaceExtensionTest,
           scalarSchema,
           enumSchema,
+          localSubscriptionSchema,
+        ],
+        typeDefs: [
+          interfaceExtensionTest,
           linkSchema,
           loneExtend,
-          localSubscriptionSchema,
           codeCoverageTypeDefs,
           schemaDirectiveTypeDefs,
         ],
@@ -1568,17 +1569,15 @@ bookingById(id: "b1") {
             },
           },
         };
-        const schema = stitchSchemas({
-          schemas: [
+        const schema = makeExecutableSchema({
+          subschemas: [
             propertySchema,
             bookingSchema,
             productSchema,
-            scalarTest,
             enumSchema,
-            linkSchema,
-            loneExtend,
             localSubscriptionSchema,
           ],
+          typeDefs: [scalarTest, linkSchema, loneExtend],
           resolvers: [
             Scalars,
             Enums,
@@ -2895,8 +2894,8 @@ fragment BookingFragment on Booking {
           },
         };
 
-        schema = stitchSchemas({
-          schemas: [schema],
+        schema = makeExecutableSchema({
+          subschemas: [schema],
           resolvers,
         });
 
@@ -2929,8 +2928,9 @@ fragment BookingFragment on Booking {
           },
         };
 
-        const schema = stitchSchemas({
-          schemas: [propertySchema, typeDefs],
+        const schema = makeExecutableSchema({
+          subschemas: [propertySchema],
+          typeDefs,
           resolvers,
         });
 
@@ -2980,7 +2980,10 @@ fragment BookingFragment on Booking {
       };
 
       const result = await graphql(
-        stitchSchemas({ schemas: [BookSchema, AuthorSchema], resolvers }),
+        makeExecutableSchema({
+          typeDefs: [BookSchema, AuthorSchema],
+          resolvers,
+        }),
         `
           query {
             book {
@@ -3036,9 +3039,13 @@ fragment BookingFragment on Booking {
       addMocksToSchema({ schema: bookSchema });
       addMocksToSchema({ schema: movieSchema });
 
-      const stitchedSchema = stitchSchemas({
-        schemas: [bookSchema, movieSchema],
-        queryTypeName: 'RootQuery',
+      const stitchedSchema = makeExecutableSchema({
+        subschemas: [bookSchema, movieSchema],
+        typeDefs: `
+          schema {
+            query: RootQuery
+          }
+        `,
       });
 
       const result = await graphql(
