+++ src/test/stitchingFromSubschemas.test.ts
@@ -13,7 +13,11 @@
 
 import { graphql } from 'graphql';
 
-import { delegateToSchema, stitchSchemas, addMocksToSchema } from '../index';
+import {
+  delegateToSchema,
+  makeExecutableSchema,
+  addMocksToSchema,
+} from '../index';
 
 const chirpTypeDefs = `
   type Chirp {
@@ -35,8 +39,8 @@ const authorTypeDefs = `
 const schemas = {};
 const getSchema = (name: string) => schemas[name];
 
-const chirpSchema = stitchSchemas({
-  schemas: [
+const chirpSchema = makeExecutableSchema({
+  typeDefs: [
     chirpTypeDefs,
     authorTypeDefs,
     `
@@ -73,8 +77,8 @@ addMocksToSchema({
   preserveResolvers: true,
 });
 
-const authorSchema = stitchSchemas({
-  schemas: [
+const authorSchema = makeExecutableSchema({
+  typeDefs: [
     chirpTypeDefs,
     authorTypeDefs,
     `
@@ -113,8 +117,8 @@ addMocksToSchema({
 schemas['chirpSchema'] = chirpSchema;
 schemas['authorSchema'] = authorSchema;
 
-const stitchedSchema = stitchSchemas({
-  schemas: Object.keys(schemas).map((schemaName) => schemas[schemaName]),
+const stitchedSchema = makeExecutableSchema({
+  subschemas: Object.keys(schemas).map((schemaName) => schemas[schemaName]),
 });
 
 describe('merging without specifying fragments', () => {
