+++ src/utils/index.ts
@@ -9,4 +9,5 @@ export { forEachField } from './forEachField';
 export { forEachDefaultValue } from './forEachDefaultValue';
 export { graphqlVersion } from './graphqlVersion';
 export { mapSchema } from './map';
+export { rewireTypes } from './rewire';
 export { default as SchemaError } from './SchemaError';
