+++ src/utils/map.ts
@@ -1,42 +1,28 @@
 import {
   GraphQLDirective,
-  GraphQLEnumType,
-  GraphQLFieldConfigArgumentMap,
-  GraphQLFieldConfigMap,
-  GraphQLInputFieldConfigMap,
-  GraphQLInputObjectType,
-  GraphQLInterfaceType,
-  GraphQLList,
   GraphQLObjectType,
-  GraphQLObjectTypeConfig,
   GraphQLNamedType,
-  GraphQLNonNull,
-  GraphQLScalarType,
   GraphQLSchema,
   GraphQLType,
-  GraphQLUnionType,
   isDirective,
   isInterfaceType,
   isEnumType,
   isInputType,
-  isInputObjectType,
-  isListType,
   isNamedType,
-  isNonNullType,
   isObjectType,
   isScalarType,
   isUnionType,
 } from 'graphql';
 
 import { toConfig } from '../polyfills/toConfig';
-import { isSpecifiedScalarType } from '../polyfills/isSpecifiedScalarType';
-import { graphqlVersion } from '../utils/graphqlVersion';
 import {
   SchemaMapper,
   MapperKind,
   NamedTypeMapper,
   DirectiveMapper,
 } from '../Interfaces';
+
+import { rewireTypes } from './rewire';
 
 export function mapSchema(
   schema: GraphQLSchema,
@@ -187,240 +173,4 @@ function getMapper(
     const directiveMapper = schemaMapper[MapperKind.DIRECTIVE];
     return directiveMapper != null ? directiveMapper : null;
   }
-}
-
-export function rewireTypes(
-  originalTypeMap: Record<string, GraphQLNamedType | null>,
-  directives: ReadonlyArray<GraphQLDirective>,
-): {
-  typeMap: Record<string, GraphQLNamedType>;
-  directives: Array<GraphQLDirective>;
-} {
-  const newTypeMap: Record<string, GraphQLNamedType> = Object.create(null);
-
-  Object.keys(originalTypeMap).forEach((typeName) => {
-    const namedType = originalTypeMap[typeName];
-
-    if (namedType == null || typeName.startsWith('__')) {
-      return;
-    }
-
-    const newName = namedType.name;
-    if (newName.startsWith('__')) {
-      return;
-    }
-
-    if (newTypeMap[newName] != null) {
-      throw new Error(`Duplicate schema type name ${newName}`);
-    }
-
-    newTypeMap[newName] = namedType;
-  });
-
-  Object.keys(newTypeMap).forEach((typeName) => {
-    newTypeMap[typeName] = rewireNamedType(newTypeMap[typeName]);
-  });
-
-  const newDirectives = directives.map((directive) =>
-    rewireDirective(directive),
-  );
-
-  return pruneTypes(newTypeMap, newDirectives);
-
-  function rewireDirective(directive: GraphQLDirective): GraphQLDirective {
-    const directiveConfig = toConfig(directive);
-    directiveConfig.args = rewireArgs(directiveConfig.args);
-    return new GraphQLDirective(directiveConfig);
-  }
-
-  function rewireArgs(
-    args: GraphQLFieldConfigArgumentMap,
-  ): GraphQLFieldConfigArgumentMap {
-    const rewiredArgs = {};
-    Object.keys(args).forEach((argName) => {
-      const arg = args[argName];
-      const rewiredArgType = rewireType(arg.type);
-      if (rewiredArgType != null) {
-        arg.type = rewiredArgType;
-        rewiredArgs[argName] = arg;
-      }
-    });
-    return rewiredArgs;
-  }
-
-  function rewireNamedType<T extends GraphQLNamedType>(type: T) {
-    if (isObjectType(type)) {
-      const config = toConfig(type);
-      const newConfig = {
-        ...config,
-        fields: () => rewireFields(config.fields),
-        interfaces: () => rewireNamedTypes(config.interfaces),
-      };
-      return new GraphQLObjectType(newConfig);
-    } else if (isInterfaceType(type)) {
-      const config = toConfig(type);
-      const newConfig = {
-        ...config,
-        fields: () => rewireFields(config.fields),
-      };
-      if (graphqlVersion() >= 15) {
-        ((newConfig as unknown) as GraphQLObjectTypeConfig<
-          any,
-          any
-        >).interfaces = () => rewireNamedTypes(config.interfaces);
-      }
-      return new GraphQLInterfaceType(newConfig);
-    } else if (isUnionType(type)) {
-      const config = toConfig(type);
-      const newConfig = {
-        ...config,
-        types: () => rewireNamedTypes(config.types),
-      };
-      return new GraphQLUnionType(newConfig);
-    } else if (isInputObjectType(type)) {
-      const config = toConfig(type);
-      const newConfig = {
-        ...config,
-        fields: () => rewireInputFields(config.fields),
-      };
-      return new GraphQLInputObjectType(newConfig);
-    } else if (isEnumType(type)) {
-      const enumConfig = toConfig(type);
-      return new GraphQLEnumType(enumConfig);
-    } else if (isScalarType(type)) {
-      if (isSpecifiedScalarType(type)) {
-        return type;
-      }
-      const scalarConfig = toConfig(type);
-      return new GraphQLScalarType(scalarConfig);
-    }
-
-    throw new Error(`Unexpected schema type: ${(type as unknown) as string}`);
-  }
-
-  function rewireFields(
-    fields: GraphQLFieldConfigMap<any, any>,
-  ): GraphQLFieldConfigMap<any, any> {
-    const rewiredFields = {};
-    Object.keys(fields).forEach((fieldName) => {
-      const field = fields[fieldName];
-      const rewiredFieldType = rewireType(field.type);
-      if (rewiredFieldType != null) {
-        field.type = rewiredFieldType;
-        field.args = rewireArgs(field.args);
-        rewiredFields[fieldName] = field;
-      }
-    });
-    return rewiredFields;
-  }
-
-  function rewireInputFields(
-    fields: GraphQLInputFieldConfigMap,
-  ): GraphQLInputFieldConfigMap {
-    const rewiredFields = {};
-    Object.keys(fields).forEach((fieldName) => {
-      const field = fields[fieldName];
-      const rewiredFieldType = rewireType(field.type);
-      if (rewiredFieldType != null) {
-        field.type = rewiredFieldType;
-        rewiredFields[fieldName] = field;
-      }
-    });
-    return rewiredFields;
-  }
-
-  function rewireNamedTypes<T extends GraphQLNamedType>(
-    namedTypes: Array<T>,
-  ): Array<T> {
-    const rewiredTypes: Array<T> = [];
-    namedTypes.forEach((namedType) => {
-      const rewiredType = rewireType(namedType);
-      if (rewiredType != null) {
-        rewiredTypes.push(rewiredType);
-      }
-    });
-    return rewiredTypes;
-  }
-
-  function rewireType<T extends GraphQLType>(type: T): T | null {
-    if (isListType(type)) {
-      const rewiredType = rewireType(type.ofType);
-      return rewiredType != null ? (new GraphQLList(rewiredType) as T) : null;
-    } else if (isNonNullType(type)) {
-      const rewiredType = rewireType(type.ofType);
-      return rewiredType != null
-        ? (new GraphQLNonNull(rewiredType) as T)
-        : null;
-    } else if (isNamedType(type)) {
-      const originalType = originalTypeMap[type.name];
-      return originalType != null ? (newTypeMap[originalType.name] as T) : null;
-    }
-
-    return null;
-  }
-}
-
-function pruneTypes(
-  typeMap: Record<string, GraphQLNamedType>,
-  directives: Array<GraphQLDirective>,
-): {
-  typeMap: Record<string, GraphQLNamedType>;
-  directives: Array<GraphQLDirective>;
-} {
-  const newTypeMap = {};
-
-  const implementedInterfaces = {};
-  Object.keys(typeMap).forEach((typeName) => {
-    const namedType = typeMap[typeName];
-
-    if (
-      isObjectType(namedType) ||
-      (graphqlVersion() >= 15 && isInterfaceType(namedType))
-    ) {
-      ((namedType as unknown) as GraphQLObjectType)
-        .getInterfaces()
-        .forEach((iface) => {
-          implementedInterfaces[iface.name] = true;
-        });
-    }
-  });
-
-  let prunedTypeMap = false;
-  const typeNames = Object.keys(typeMap);
-  for (let i = 0; i < typeNames.length; i++) {
-    const typeName = typeNames[i];
-    const type = typeMap[typeName];
-    if (isObjectType(type) || isInputObjectType(type)) {
-      // prune types with no fields
-      if (Object.keys(type.getFields()).length) {
-        newTypeMap[typeName] = type;
-      } else {
-        prunedTypeMap = true;
-      }
-    } else if (isUnionType(type)) {
-      // prune unions without underlying types
-      if (type.getTypes().length) {
-        newTypeMap[typeName] = type;
-      } else {
-        prunedTypeMap = true;
-      }
-    } else if (isInterfaceType(type)) {
-      // prune interfaces without fields or without implementations
-      if (
-        Object.keys(type.getFields()).length &&
-        implementedInterfaces[type.name]
-      ) {
-        newTypeMap[typeName] = type;
-      } else {
-        prunedTypeMap = true;
-      }
-    } else {
-      newTypeMap[typeName] = type;
-    }
-  }
-
-  // every prune requires another round of healing
-  return prunedTypeMap
-    ? rewireTypes(newTypeMap, directives)
-    : { typeMap, directives };
 }
