+++ src/wrap/wrapSchema.ts
@@ -16,7 +16,6 @@ import {
 import { toConfig } from '../polyfills';
 import { defaultMergedResolver } from '../delegate';
 import { mapSchema } from '../utils';
-import resolveFromParentTypename from '../utils/resolveFromParentTypename';
 import { applySchemaTransforms } from '../utils/transforms';
 
 import { generateProxyingResolvers } from './generateProxyingResolvers';
@@ -85,12 +84,12 @@ function createWrappingSchema(
     },
     [MapperKind.INTERFACE_TYPE]: (type) => {
       const config = toConfig(type);
-      config.resolveType = (parent) => resolveFromParentTypename(parent);
+      delete config.resolveType;
       return new GraphQLInterfaceType(config);
     },
     [MapperKind.UNION_TYPE]: (type) => {
       const config = toConfig(type);
-      config.resolveType = (parent) => resolveFromParentTypename(parent);
+      delete config.resolveType;
       return new GraphQLUnionType(config);
     },
   });
